+++ src/routes/+layout.svelte
@@ -17,6 +17,8 @@
 
   import { completeFilter, ignoredFilter, incompleteFilter, nameFilter, oneRegionFilter } from "../store.js";
 
+  export const ssr = false;
+
   let navbarIsOpen = false;
 
   function handleNavbarUpdate(event) {